If check.chk exists
The submit script will look for the check file to copy if it exists. If it doesn't then it will not output an error and continue as usual

update

+++ Servers/Zeus/ARC/.arc/submit.py
@@ -40,7 +40,11 @@ g09 < input.gjf > input.log
 
 cp input.* "$PBS_O_WORKDIR/"
 
-cp check.* "$PBS_O_WORKDIR/"
+if [ ! -f check.* ]; then
+	:
+else
+	cp check.* "$PBS_O_WORKDIR/"
+fi
 
 rm -vrf $GAUSS_SCRDIR
